requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
ckanapi==4.1
click==6.7
docopt==0.6.2
idna==2.6
python-dotenv==0.7.1
requests==2.18.4
urllib3==1.22
